requirements.txt
matplotlib==3.5.0
numpy==1.21.2
opencv_python==4.5.4.60
Pillow==9.1.0
scipy==1.7.3
torch==1.8.0
torchvision==0.9.0
tqdm==4.62.3
